pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-export"
version = "1.2.0"
description = "Flexible Telegram data export tool"
license = "MIT"
requires-python = ">=3.11"
readme = "README.md"
authors = [{name = "Vitaly Ostanin", email = "vitaly.ostanin@gmail.com"}]
dependencies = [
    "telethon>=1.36",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "jinja2>=3.1",
    "click>=8.0",
    "rich>=15.0",
]

[project.urls]
Homepage = "https://github.com/VitalyOstanin/tg-export"
Repository = "https://github.com/VitalyOstanin/tg-export"
Issues = "https://github.com/VitalyOstanin/tg-export/issues"

[project.optional-dependencies]
proxy = [
    "python-socks[asyncio]>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "pyright>=1.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = [
    "error::DeprecationWarning",
    "ignore::DeprecationWarning:telethon",
]
timeout = 60

[tool.ruff]
target-version = "py311"
line-length = 110
extend-exclude = ["export_output", "docs/reviews"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in default arg (used by Click)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.pyright]
include = ["tg_export", "tests"]
exclude = ["export_output", "docs/reviews", ".venv"]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportOptionalMemberAccess = "warning"

[tool.setuptools.packages.find]
include = ["tg_export*"]

[project.scripts]
tg-export = "tg_export.cli:main"
